+++ src/Webserver.API/Services/IApiServiceFactory.cs
@@ -2,9 +2,10 @@
 using System.Threading.Tasks;
 using Siemens.Simatic.S7.Webserver.API.Models;
 using Siemens.Simatic.S7.Webserver.API.Services.PlcProgram;
+using Siemens.Simatic.S7.Webserver.API.Services.RequestHandling;
 using Siemens.Simatic.S7.Webserver.API.Services.WebApp;
 
-namespace Siemens.Simatic.S7.Webserver.API.Services.RequestHandling
+namespace Siemens.Simatic.S7.Webserver.API.Services
 {
     /// <summary>
     /// Factory to create services
